+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/BubbleGraph/MultipleDataset.test.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import BubbleMultipleDataset from '@cerner/terra-graphs/lib/components/Bubble/BubbleMultipleDataset';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import getSimpleAxisData from '@cerner/terra-graphs-docs/src/example-datasets/graphConfigObjects/Bubble/simpleAxisData';
+import bubbleDataBasic from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataBasic.js';
+import bubbleDataCustomRadius from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataCustomRadius.js';
+
+const graphConfig = utils.deepClone(getSimpleAxisData('#multipleDatasetBubbleGraph'));
+const dataset1 = utils.deepClone(bubbleDataBasic);
+const dataset2 = utils.deepClone(bubbleDataCustomRadius);
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <BubbleMultipleDataset graphID="multipleDatasetBubbleGraph" graphConfig={graphConfig} dataset={[dataset1, dataset2]} />
+  </React.Fragment>
+);