+++ packages/carbon-graphs/CHANGELOG.md
@@ -8,6 +8,7 @@
   * Fixed reflow bug in timeline triggered by `showLegend=false`.
   * Fixed retaining legend state in paired result graph during reflow.
   * Fixed isolation of different graph instances in same page.
+  * Allowed Bubble constants to be accessible through the `Carbon.helpers` object.
 
 ## 2.17.0 - (April 27, 2021)