+++ packages/terra-graphs/CHANGELOG.md
@@ -1,11 +1,10 @@
 # ChangeLog
 
 ## Unreleased
-
-* Changed
-  * Create Components, examples and WDIO test cases for Line, Bar and Paired Result graph types.
   
 * Added
+  * Created components, examples and WDIO test cases for Line, Bar and Paired Result graph types.
+  * Added bubble components, wdio tests and corresponding examples.
   * Added documentation for `reflowMultipleDatasets`.
   * Created Component and WDIO test cases for Scatter graph types.